backend/pyproject.toml
[tool.poetry]
name = "lavarrock"
version = "0.1.0"
description = "Self-hosted Obsidian-style AI copilot editor"
authors = ["Nick Hulley <contact@nick-hulley.com>"]
readme = "README.md"
packages = [{include = "lavarrock"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = "^0.24.0"
python-dotenv = "^1.0.0"
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
ollama = "^0.1.0"
boto3 = "^1.28.0"
sqlalchemy = "^2.0.0"
alembic = "^1.12.0"
websockets = "^12.0"
aiosqlite = "^0.19.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
mypy = "^1.7.0"
ruff = "^0.1.0"
pytest-cov = "^4.1.0"
black = "^23.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
